requirements.txt
requests==2.31.0
google-cloud-secretmanager==2.19.0
google-cloud-bigquery==3.21.0 